pyproject.toml
[tool.poetry]
name = "maya_sawa"
version = "0.1.0"
description = "A powerful document Q&A system based on FastAPI, LangChain, and ChromaDB"
authors = ["TYK <tianyikao@gmail.com>"]
readme = "README.md"
packages = [{include = "maya_sawa"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.9"
uvicorn = "^0.27.1"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
langchain = "0.3.25"
langchain-openai = "0.3.23"
langchain-community = "^0.3.25"
psycopg2-binary = "^2.9.9"
httpx = "^0.27.0"
redis = "^5.0.1"
openai = "^1.0.0"
# Celery for async task processing
celery = "^5.4.0"
kombu = "^5.4.0"
# Multi AI providers
google-generativeai = "^0.8.3"
dashscope = "^1.14.0"
# SQLAlchemy for paprika database connection
sqlalchemy = "^2.0.0"
aiosqlite = "^0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "*"
black = "*"

[tool.poetry.scripts]
start = "maya_sawa.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
